requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0
albumentations==1.3.1

# Deep Learning
torch==2.1.1
torchvision==0.16.1
timm==0.9.12
transformers==4.36.0
onnx==1.15.0
onnxruntime==1.16.3

# Data science
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
structlog==23.2.0

# Monitoring & Testing
prometheus-client==0.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1